% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_query_so.R
\name{format_query_so}
\alias{format_query_so}
\title{Format input H5 as a Seurat object}
\usage{
format_query_so(in_h5, allow_few_cells = FALSE, dup_pattern = "__dup__")
}
\arguments{
\item{in_h5}{Filepath to input h5 file}

\item{allow_few_cells}{Logical value, default FALSE. If TRUE allows proceeding when
input dataset has fewer thatn 200 cells but bootstrapping up to >=200 cells. If FALSE,
errors out if too few cells}

\item{dup_pattern}{String value prefix for bootstrapped cellnames when allow_few_cells
behavior is triggered. Default '__dup__'}
}
\description{
Formats the .H5 file as a Seurat object with option to
bootstrap if there are fewer than 200 cells.
}
\details{
For bootstrapping, the entire dataset will be duplicated 
enough times to obtain 200 cells prior to labeling. Duplicates
will be removed downstream.
}
