% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_reference.R
\name{read_reference}
\alias{read_reference}
\title{Read in and validate input reference object file}
\usage{
read_reference(ref_fp)
}
\arguments{
\item{ref_fp}{File path to reference object, an .rds or .h5seurat object}
}
\value{
Seurat object contained in input file
}
\description{
Currently supports .rds or .h5seurat input file and enforces read in
object as type 'Seurat'.
}
