% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_pbmc_10k_v3}
\alias{variable_pbmc_10k_v3}
\title{Seurat reference 10X PBMC dataset}
\format{A Seurat object with data from 5,000 genes and 11,537 cells}
\usage{
variable_pbmc_10k_v3
}
\description{
This dataset was retrieved from 10x Genomics here: http://cf.10xgenomics.com/samples/cell-exp/3.0.0/pbmc_10k_v3/pbmc_10k_v3_filtered_feature_bc_matrix.h5
}
\details{
It was loaded as a Seurat object, normalized with default settings, and 5,000 variable features were selected.

The data was then filtered to retain only these variable features for portability.

The script used to generate this file is available within this package in inst/make_reference_data.R
}
\keyword{datasets}
