% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_pbmc_10k_v3}
\alias{variable_pbmc_10k_v3}
\title{Seurat reference 10X PBMC dataset}
\format{A Seurat object with data from 5,000 genes and 11,537 cells}
\usage{
variable_pbmc_10k_v3
}
\description{
This dataset was retrieved from the Seurat website from this URL: https://www.dropbox.com/s/3f3p5nxrn5b3y4y/pbmc_10k_v3.rds?dl=1
}
\details{
It was loaded and 5,000 variable features were selected.

The data was then filtered to retain only these variable features for portability.

The script used to generate this file is available within this package in inst/make_reference_data.R
}
\keyword{datasets}
